+++ DTCD-ExtensionCommonPrimitives/src/primitives/index.js
@@ -3,11 +3,16 @@ import SimpleEdge from './SimpleEdge';
 // import DashedEdge from './DashedEdge';
 // import TargetNode from './TargetNode';
 // import RiskNode from './RiskNode';
-import GroupNode from './GroupNode/GroupNode';
+// import GroupNode from './GroupNode/GroupNode';
 // import MathOperationNode from './MathOperationNode';
 // import InPort from './InPort';
 // import OutPort from './OutPort';
-import HTMLNode from './HTMLNode/HTMLNode';
-import ChartNode from './ChartNode/ChartNode';
+// import HTMLNode from './HTMLNode/HTMLNode';
+// import ChartNode from './ChartNode/ChartNode';
 
-export default [SimpleEdge, GroupNode, HTMLNode, ChartNode];
+export default [
+    SimpleEdge, 
+    // GroupNode, 
+    // HTMLNode, 
+    // ChartNode
+];